# resources/docs/swagger.yaml
definitions:
...
  request.RoleCreateRequest:
    properties:
      dataScope:
        description: 数据范围
        type: string
      deptIds:
        description: 部门编号数组
        items:
          type: integer
        type: array
      menuIds:
        description: 菜单编号数据
        items:
          type: integer
        type: array
      orderNum:
        description: 显示顺序
        minimum: 0
        type: integer
      remark:
        description: 备注
        type: string
      roleKey:
        description: 角色权限字符串
        type: string
      roleName:
        description: 角色名称
        type: string
    required:
    - dataScope
    - roleKey
    - roleName
    type: object
  request.RoleUpdateRequest:
    properties:
      dataScope:
        description: 数据范围
        type: string
      deptIds:
        description: 部门编号数组
        items:
          type: integer
        type: array
      id:
        description: 角色ID
        type: integer
      menuIds:
        description: 菜单编号数据
        items:
          type: integer
        type: array
      orderNum:
        description: 显示顺序
        minimum: 0
        type: integer
      remark:
        description: 备注
        type: string
      roleKey:
        description: 角色权限字符串
        type: string
      roleName:
        description: 角色名称
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
    required:
    - dataScope
    - id
    - roleKey
    - roleName
    type: object
...
  request.UserUpdateRequest:
    properties:
      PostIds:
        description: 岗位编号数组
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      avatar:
        description: 头像地址
        type: string
      deptID:
        description: 部门ID
        type: integer
      email:
        description: 用户邮箱
        type: string
      id:
        description: 用户ID
        type: integer
      mobile:
        description: 手机号码
        type: string
      realName:
        description: 用户昵称
        type: string
      remark:
        description: 备注
        type: string
      roleIds:
        description: 角色编号数组
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      sex:
        description: 用户性别（0男 1女 2未知）
        enum:
        - "0"
        - ' 1'
        - ' 2'
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
    required:
    - avatar
    - deptID
    - email
    - id
    - mobile
    - realName
    - remark
    - sex
    type: object
...
  response.DeptResponse:
    properties:
      ancestors:
        description: 祖级列表
        type: string
      children:
        description: 字部门列表
        items:
          $ref: '#/definitions/response.DeptResponse'
        type: array
      childrenSize:
        description: 子部门数量
        type: integer
      createBy:
        description: 创建者
        type: string
      createTime:
        description: 创建时间
        type: string
      delFlag:
        description: 删除标志（0代表存在 2代表删除）
        type: string
      deptCode:
        description: 部门编码
        type: string
      deptName:
        description: 部门名称
        type: string
      id:
        description: 部门id
        type: integer
      orderNum:
        description: 显示顺序
        type: integer
      parentID:
        description: 父部门id
        type: integer
      remark:
        description: 备注
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
      updateBy:
        description: 更新者
        type: string
      updateTime:
        description: 更新时间
        type: string
    type: object
...
  response.MenuResponse:
    properties:
      children:
        description: 子菜单
        items:
          $ref: '#/definitions/response.MenuResponse'
        type: array
      component:
        description: 组件路径
        type: string
      createBy:
        description: 创建者
        type: string
      createTime:
        description: 创建时间
        type: string
      delFlag:
        description: 删除标志（0代表存在 2代表删除）
        type: string
      icon:
        description: 菜单图标
        type: string
      id:
        description: 菜单ID
        type: integer
      isCache:
        description: 是否缓存（0缓存 1不缓存）
        type: integer
      isFrame:
        description: 是否为外链（0是 1否）
        type: integer
      menuKey:
        description: 菜单标识
        type: string
      menuName:
        description: 菜单名称
        type: string
      menuType:
        description: 菜单类型（D目录 M菜单 A按钮）
        type: string
      orderNum:
        description: 显示顺序
        type: integer
      parentID:
        description: 父菜单ID
        type: integer
      path:
        description: 路由地址
        type: string
      remark:
        description: 备注
        type: string
      requestMethod:
        description: 请求方法
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
      updateBy:
        description: 更新者
        type: string
      updateTime:
        description: 更新时间
        type: string
      visible:
        description: 菜单状态（0显示 1隐藏）
        type: string
    type: object
...
  response.PostResponse:
    properties:
      createBy:
        description: 创建者
        type: string
      createTime:
        description: 创建时间
        type: string
      delFlag:
        description: 删除标志（0代表存在 2代表删除）
        type: string
      id:
        description: 岗位ID
        type: integer
      orderNum:
        description: 显示顺序
        type: integer
      postName:
        description: 岗位名称
        type: string
      remark:
        description: 备注
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
      updateBy:
        description: 更新者
        type: string
      updateTime:
        description: 更新时间
        type: string
    type: object
  response.RestRes:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: status ok
        type: string
    type: object
  response.RoleResponse:
    properties:
      createBy:
        description: 创建者
        type: string
      createTime:
        description: 创建时间
        type: string
      dataScope:
        description: 数据范围（1：全部数据权限 2：自定数据权限 3：本部门数据权限 4：本部门及以下数据权限）
        type: string
      delFlag:
        description: 删除标志（0代表存在 2代表删除）
        type: string
      deptList:
        description: 部门列表
        items:
          $ref: '#/definitions/response.DeptResponse'
        type: array
      id:
        description: 角色ID
        type: integer
      menuList:
        description: 菜单列表
        items:
          $ref: '#/definitions/response.MenuResponse'
        type: array
      orderNum:
        description: 显示顺序
        type: integer
      remark:
        description: 备注
        type: string
      roleKey:
        description: 角色权限字符串
        type: string
      roleName:
        description: 角色名称
        type: string
      status:
        description: 帐号状态（0正常 1停用）
        type: string
      updateBy:
        description: 更新者
        type: string
      updateTime:
        description: 更新时间
        type: string
    type: object
  response.SessionResponse:
    properties:
      clientIp:
        description: 请求地址
        type: string
      createAt:
        description: 创建时间
        type: string
      expiresAt:
        description: 是否阻止
        type: string
      id:
        description: 唯一标识
        type: string
      isBlocked:
        description: 是否阻止
        type: boolean
      realName:
        description: 真实姓名
        type: string
      refreshToken:
        description: 刷新秘钥
        type: string
      remark:
        description: 备注
        type: string
      userAgent:
        description: 请求信息
        type: string
      userName:
        description: 用户名
        type: string
    type: object
...
  response.UserResponse:
    properties:
      avatar:
        description: 头像地址
        type: string
      createBy:
        description: 创建者
        type: string
      createTime:
        description: 创建时间
        type: string
      delFlag:
        description: 删除标志（0代表存在 2代表删除）
        type: string
      dept:
        $ref: '#/definitions/response.DeptResponse'
        description: 部门信息
      deptID:
        description: 部门ID
        type: integer
      deptName:
        description: 部门名称
        type: string
      email:
        description: 用户邮箱
        type: string
      id:
        description: 用户ID
        type: integer
      menuList:
        description: 菜单信息
        items:
          $ref: '#/definitions/response.MenuResponse'
        type: array
      mobile:
        description: 手机号码
        type: string
      postList:
        description: 岗位信息
        items:
          $ref: '#/definitions/response.PostResponse'
        type: array
      posts:
        description: 岗位编号数组
        type: string
      realName:
        description: 用户昵称
        type: string
      remark:
        description: 备注
        type: string
      roleList:
        description: 角色信息
        items:
          $ref: '#/definitions/response.RoleResponse'
        type: array
      roles:
        description: 角色信息
        items:
          type: string
        type: array
      sex:
        description: 用户性别（0男 1女 2未知）
        type: string
      status:
        description: 帐号状态（0正常 1停用）
        type: string
      updateBy:
        description: 更新者
        type: string
      updateTime:
        description: 更新时间
        type: string
      userName:
        description: 用户账号
        type: string
    type: object
info:
  contact: {}
  description: This is a sample golang project
  title: Admin GOGOGO
  version: 0.0.1
paths:
...
  /post/p:
    post:
      consumes:
      - application/json
      description: 获取所有操作日志信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.SessionResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 操作日志列表查询
      tags:
      - 系统操作日志
  /role:
    post:
      consumes:
      - application/json
      description: 新增角色信息
      parameters:
      - description: 新增角色信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.RoleCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 角色新增
      tags:
      - 系统角色
    put:
      consumes:
      - application/json
      description: 更新角色信息
      parameters:
      - description: 更新角色信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.RoleUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 角色更新
      tags:
      - 系统角色
  /role/:id:
    delete:
      consumes:
      - application/json
      description: 伪删除角色信息
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 角色伪删除
      tags:
      - 系统角色
    get:
      consumes:
      - application/json
      description: 详情角色信息
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.RoleResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 角色详情
      tags:
      - 系统角色
  /role/list:
    get:
      consumes:
      - application/json
      description: 获取所有角色信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.RoleResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 角色列表查询
      tags:
      - 系统角色
  /role/p:
    post:
      consumes:
      - application/json
      description: 获取所有角色信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.RoleResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 角色列表查询
      tags:
      - 系统角色
  /user:
    post:
      consumes:
      - application/json
      description: 新增用户信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 创建用户
      tags:
      - 系统用户
    put:
      consumes:
      - application/json
      description: 更新用户信息
      parameters:
      - description: 更新用户信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 更新用户
      tags:
      - 系统用户
  /user/:id:
    delete:
      consumes:
      - application/json
      description: 删除用户信息
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 删除用户
      tags:
      - 系统用户
    get:
      consumes:
      - application/json
      description: 用户信息详情（关联部门、角色、岗位）
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 用户详情（关联部门、角色、岗位）
      tags:
      - 系统用户
...
  /user/p:
    get:
      consumes:
      - application/json
      description: 分页获取所有用户信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.UserResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 用户分页查询
      tags:
      - 系统用户
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"